agent/src/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.28.1
databricks-sdk==0.40.0
openai==1.55.3
pydantic==2.10.3
python-dateutil==2.9.0
